requirements.txt
qrcode==6.1
Pillow==7.2.0